+++ src/components/layer-controller/shared-channel-controls.js
@@ -6,13 +6,13 @@ import Select from '@material-ui/core/Select';
 /**
  * Dropdown for selecting a channel.
  * @prop {function} handleChange Callback for each new selection.
- * @prop {boolean} disableOptions Whether or not to allow options.
  * @prop {array} channelOptions List of available selections, like ['DAPI', 'FITC', ...].
+ * @prop {boolean} disabled Whether or not the component is disabled.
  * @prop {number} selectionIndex Current numeric index of a selection.
  */
 export function ChannelSelectionDropdown({
   handleChange,
-  disableOptions,
+  disabled,
   channelOptions,
   selectionIndex,
 }) {
@@ -23,7 +23,7 @@ export function ChannelSelectionDropdown({
       onChange={e => handleChange(Number(e.target.value))}
     >
       {channelOptions.map((opt, i) => (
-        <option disabled={disableOptions} key={opt} value={i}>
+        <option disabled={disabled} key={opt} value={i}>
           {opt}
         </option>
       ))}
@@ -35,13 +35,17 @@ export function ChannelSelectionDropdown({
  * Checkbox for toggling on/off of a channel.
  * @prop {string} color Current color for this channel.
  * @prop {boolean} checked Whether or not this channel is "on".
+ * @prop {boolean} disabled Whether or not the component is disabled.
  * @prop {function} toggle Callback for toggling on/off.
  */
-export function ChannelVisibilityCheckbox({ color, checked, toggle }) {
+export function ChannelVisibilityCheckbox({
+  color, checked, toggle, disabled,
+}) {
   return (
     <Checkbox
       onChange={toggle}
       checked={checked}
+      disabled={disabled}
       style={{ color, '&$checked': { color } }}
     />
   );