+++ src/components/spatial/Spatial.js
@@ -237,8 +237,9 @@ class Spatial extends AbstractSpatialOrScatterplot {
 
   createScaleBarLayer() {
     const {
-      viewState, width, height, imageLayerLoaders = {},
+      viewState, width, height, imageLayerLoaders = {}, layers,
     } = this.props;
+    const use3d = (layers || []).some(i => i.use3d);
     // Just get the first layer/loader since they should all be spatially
     // resolved and therefore have the same unit size scale.
     const loaders = Object.values(imageLayerLoaders);
@@ -248,7 +249,7 @@ class Spatial extends AbstractSpatialOrScatterplot {
     const source = getSourceFromLoader(loader);
     if (!source.meta) return null;
     const { physicalSizes } = source.meta;
-    if (physicalSizes) {
+    if (physicalSizes && !use3d) {
       const { x } = physicalSizes;
       const { unit, size } = x;
       if (unit && size) {
@@ -291,6 +292,12 @@ class Spatial extends AbstractSpatialOrScatterplot {
       transparentColor: layerDef.transparentColor,
       colors: layerDef.channels.map(c => c.color),
       sliders: layerDef.channels.map(c => c.slider),
+      resolution: layerDef.resolution,
+      renderingMode: layerDef.renderingMode,
+      xSlice: layerDef.xSlice,
+      ySlice: layerDef.ySlice,
+      zSlice: layerDef.zSlice,
+      callback: layerDef.callback,
       visibilities: layerDef.channels.map(
         c => (!layerDef.visible && typeof layerDef.visible === 'boolean' ? false : c.visible),
       ),
@@ -326,15 +333,19 @@ class Spatial extends AbstractSpatialOrScatterplot {
         cellColorHeight: this.color.height,
         cellColorWidth: this.color.width,
         excludeBackground: true,
+        onViewportLoad: layerProps.callback,
       });
     }
     const [Layer, layerLoader] = getLayerLoaderTuple(
       data,
-      rawLayerDef.isBitmask,
+      layerDef.use3d,
     );
+
     return new Layer({
       loader: layerLoader,
-      id: `image-layer-${layerDef.index}-${i}`,
+      id: `${layerDef.use3d ? 'volume' : 'image'}-layer-${
+        layerDef.index
+      }-${i}`,
       colorValues: layerProps.colors,
       sliderValues: layerProps.sliders,
       loaderSelection,
@@ -343,15 +354,26 @@ class Spatial extends AbstractSpatialOrScatterplot {
       colormap: layerProps.colormap,
       modelMatrix,
       transparentColor: layerProps.transparentColor,
+      resolution: layerProps.resolution,
+      renderingMode: layerProps.renderingMode,
+      pickable: false,
+      xSlice: layerProps.xSlice,
+      ySlice: layerProps.ySlice,
+      zSlice: layerProps.zSlice,
+      onViewportLoad: layerProps.callback,
     });
   }
 
   createImageLayers() {
-    const { layers, imageLayerLoaders = {} } = this.props;
+    const { layers, imageLayerLoaders = {}, rasterLayersCallbacks = [] } = this.props;
+    const use3d = (layers || []).some(i => i.use3d);
+    const use3dIndex = (layers || []).findIndex(i => i.use3d);
     return (layers || [])
       .filter(layer => (layer.type === 'raster' || layer.type === 'bitmask'))
+      .filter(layer => (use3d ? layer.use3d === use3d : true))
       .map((layer, i) => this.createImageLayer(
-        layer, imageLayerLoaders[layer.index], i,
+        { ...layer, callback: rasterLayersCallbacks[use3d ? use3dIndex : i] },
+        imageLayerLoaders[layer.index], i,
       ));
   }
 
@@ -518,7 +540,7 @@ class Spatial extends AbstractSpatialOrScatterplot {
       this.forceUpdate();
     }
 
-    if (['layers', 'imageLayerLoaders', 'cellColors', 'cellHighlight'].some(shallowDiff)) {
+    if (['layers', 'imageLayerLoaders', 'cellColors', 'cellHighlight', 'rasterLayersCallbacks'].some(shallowDiff)) {
       // Image layers changed.
       this.onUpdateImages();
       this.forceUpdate();