+++ src/components/spatial/constants.js
@@ -1,3 +1,5 @@
+import { RENDERING_MODES } from '@hms-dbmi/viv';
+
 export const GLOBAL_LABELS = ['z', 't'];
 
 export const DEFAULT_RASTER_DOMAIN_TYPE = 'Min/Max';
@@ -8,6 +10,8 @@ export const DEFAULT_RASTER_LAYER_PROPS = {
   opacity: 1,
   domainType: DEFAULT_RASTER_DOMAIN_TYPE,
   transparentColor: [0, 0, 0],
+  renderingMode: RENDERING_MODES.ADDITIVE,
+  use3d: false,
 };
 
 export const DEFAULT_MOLECULES_LAYER = {